+++ app/api/v1/data/sunrise_sunset.json
@@ -1,0 +1,4 @@
+{
+    "sunrise": "04:45 AM",
+    "sunset": "09:22 PM"
+}
