+++ app/data/sunrise_sunset.json
@@ -1,0 +1,4 @@
+{
+    "sunrise": "05:49 AM",
+    "sunset": "09:58 PM"
+}
